corrections in service, app

+++ service.py
@@ -51,7 +51,7 @@ def remove_bron(connect, pr, title, author):
         return "Книга не найдена"
     book_id = book_result[0]
 
-    cursor.execute("""SELECT pr, id FROM holds
+    cursor.execute("""SELECT pr, book_id FROM holds
                    WHERE pr = ? and id = ?""", (pr, book_id))
     check_bron = cursor.fetchone()
     if not check_bron:
@@ -78,7 +78,7 @@ def take_book(connect, pr, title, author):
     free, book_id = book_result
 
     cursor.execute("""SELECT pr, book_id FROM holds
-                   WHERE pr = ? and id = ?""", (pr, book_id))
+                   WHERE pr = ? and book_id = ?""", (pr, book_id))
     check_bron = cursor.fetchone()
 
     cursor.execute("""SELECT count(book_id) FROM loans
@@ -167,22 +167,32 @@ def get_overdue_books(connect):
     cursor.execute("""SELECT r.pr, r.full_name, b.title, b.author, l.date 
                    FROM loans l 
                    JOIN readers r ON l.pr = r.pr 
-                   JOIN books b ON l.book_id = b.id 
-                   WHERE date(l.date, '+14 days') < date('now')""")
+                   JOIN books b ON l.book_id = b.id""")
     overdue_books = []
-    for title, author, date_loan_str in cursor.fetchall():
+    today = datetime.now().date()
+    for pr, title, author, date_loan_str in cursor.fetchall():
         date_loan = datetime.strptime(date_loan_str, "%d/%m/%y").date()
         date_return = date_loan + timedelta(days=14)
-        date_return_str = date_return.strftime("%d/%m/%y")
-        overdue_books.append((title, author, date_loan_str, date_return_str))
+        if today > date_loan:
+            date_return_str = date_return.strftime("%d/%m/%y")
+            overdue_books.append((pr, title, author, date_return_str))
     return overdue_books
 
     
 def auto_remove_reservation(connect):
     cursor = connect.cursor()
-    cursor.execute("""DELETE FROM holds 
-                   WHERE date(date, '+5 days') < date('now')""")
+    cursor.execute("""SELECT pr, book_id, date FROM holds""")
+    cnt = 0
+    today = datetime.now().date()
+    for pr, book_id, date_reserve_str in cursor.fetchall():
+        date_reserve = datetime.strptime(date_reserve_str, "%d/%m/%y").date()
+        expiry_date = date_reserve + timedelta(days=5)
+        if today > expiry_date:
+             cursor.execute("""DELETE FROM holds 
+                           WHERE pr = ? AND book_id = ?""", (pr, book_id))
+             cnt += 1
     connect.commit()
+    return f"Удалено {cnt} броней"
 
 def search_books(connect, title = None, author = None, genre = None):
     cursor = connect.cursor()
